feat: Update reservation page to include 'Recibo' button
The 'reserva.html' template was modified to replace the 'Pagar' button with a 'Recibo' button for each reservation. Clicking the 'Recibo' button generates a receipt for the reservation using the '/recibo/create/{id}' endpoint. This change enhances the functionality of the application by allowing users to easily generate receipts for their reservations.